cmsTaskOverviewWebServer: trivial refactorings
